+++ bundle/src/index.global.ts
@@ -1,4 +1,5 @@
 import * as Internal from '@fullcalendar/core/internal'
+import * as Preact from '@fullcalendar/core/preact'
 
-export { Internal }
+export { Internal, Preact }
 export * from './index.js'