+++ CHANGELOG.md
@@ -3,6 +3,7 @@ multi-month-view release
 ------------------------
 - more-link hover effect
 - post-release: fix selection across separate months
+- fix: FullCalendar.Preact exposed
 
 
 v6.0.4 (2023-01-13)